+++ net-reporter/src/continuous_aggregate/mod.rs
@@ -43,3 +43,4 @@ pub mod bandwidth_per_endpoint;
 pub mod network_bandwidth_per_protocol;
 pub mod network_bandwidth;
 pub mod network_overview_filters;
+pub mod total_http_requests;