+++ net-reporter/src/query/charts/total_http_requests/response/total_http_requests_bucket.rs
@@ -1,0 +1,19 @@
+use net_reporter_api::api::total_http_requests::http_requests_bucket::HttpRequestsBucketDTO;
+
+use sqlx::types::chrono::DateTime;
+use sqlx::types::chrono::Utc;
+
+#[derive(sqlx::FromRow, Clone, Debug)]
+pub struct TotalHttpRequestsBucketResponse {
+    bucket: DateTime<Utc>,
+    total_requests: i64,
+}
+
+impl From<TotalHttpRequestsBucketResponse> for HttpRequestsBucketDTO {
+    fn from(value: TotalHttpRequestsBucketResponse) -> Self {
+        HttpRequestsBucketDTO::new(
+            value.bucket.timestamp_millis(),
+            value.total_requests
+        )
+    }
+}